Trying out new decoder

+++ src/Thawed/IByteReader.cs
@@ -2,6 +2,6 @@ namespace Thawed
 {
     public interface IByteReader
     {
-        byte ReadOne();
+        int ReadOne();
     }
 }
